+++ changelog.md
@@ -2,7 +2,8 @@
 
 ## 1.12.0 (Future)
 
-- Added masteries! (documentation TBD)
+- Added mastery ritual! (documentation TBD)
+- Added masteries for Grandmaster, Chekhov's Gun, Pincushion, Platinum Star, and Natascha
 - Added 18 new hats
 - Fixed issue with knight pieces fired from Grandmaster getting stuck in infinite loops when no enemies are around
 - Fixed Aimu Hakurei being able to build graze off of co-op partner's bullets
@@ -10,6 +11,7 @@
 - Bulletbot Implant, Bionic Finger, and Gorgun's Eye can now spawn in Handy's shop from [Knife to a Gunfight](https://thunderstore.io/c/enter-the-gungeon/p/Skilotar/Knife_to_a_Gunfight/)!
 - Fixed visual bug with misaligned muzzle flash for Holy Water Gun 
 - Fixed bug (for the second time) where loading the secret floor after returning to the breach would cause an infinite loading screen
+- Lowered volume on Natascha spinup sound
 
 ## 1.11.5 (2024-04-20)